ImGuiDiskManipulator: transfer host to MSX
The MSXtar code to import files/directories in an MSX disk was
inconsistent. This code is used by the DiskManipulator class and thus
also by the 'diskmanipulator' Tcl command:
* Importing a file works as expected.
* But importing a directory is weird: it imports all the files (and
  recursively all sub-directories) from that host directory, but WITHOUT
  creating a corresponding top-level directory on the MSX side.

I can, sort of, understand where this is coming from: when
interactively importing single files or single directories on the
command line this can be convenient (then you don't need to know about
the Tcl 'glob' command).

But this behavior makes it harder to import a host directory structure
as is. E.g. if you import a mix a files and directories, you expect this
to be mirrored on the MSX side. But instead you'll loose the 1st level
of the host-directories. This is very annoying when used in Tcl script
or now also when trying to use this in ImGuiDiskManipulator.

This patch adds a new method in MSXtar that imports a host file/dir and
_keeps_ the top-level directory in case of a host-directory.

The rest of this patch uses this functionality in ImGuiDiskManipulator.
It also fixes an off-by-one error in the month while translation between
FAT and unix (time_t) dates.

For backwards compatibility I've kept the inconsistent behavior of the
diskmanipulator Tcl command.