ImGuiDiskManipulator: new disk image
This almost finishes the ImGuiDiskManipulator stuff.

The only remaining problem is that browsing for a name for the new disk
image doesn't work yet. The problem is that 'new disk image' opens a
modal window, and browsing for a name also opens a modal window. At the
moment this 2nd modal window closes the first one.

Dear ImGui _does_ support multiple simultaneous modal windows. But only
if they are properly nested. That's currently not the case. Fixing this
requires some refactoring in ImGuiOpenFile. I'll try to do that soon.