ImGuiOpenFile: allow to set different painters
Before this patch the open-file-dialog didn't work for
  Tools > Disk Manipultor > Create new disk image > Select new file
The reason was that, for Dear ImGui, for nested modal windows, the draw
calls must also be nested. But before this patch the open file dialog
was always drawn from the top hierarchical level.

This patch allows to set different painters for this dialog.