ImGui: popup on drag-and-drop disk-image
When you drag-and-drop a tape/rom/disk image on the main openMSX window,
we (try to) insert that image. If there's more than one disk drive or
cartridge slot present we open a popup to ask which drive/slot to use.
When there's only a single drive/slot we immediately use that one,
without opening a popup.

This patch changes that: you now always get visual feedback that the
insert was successful (or not). E.g. even if only one drive is present,
after drag-and-drop a disk image you get a popup:
  Inserted mydisk.dsk in disk drive A.

At the moment this popup only closes on the next ImGui interaction.
Typically _any_ mouse-click (anywhere). We could automatically close the
popup, e.g. after 5 seconds or so (but for some this will be too long or
too short). But maybe this is already good enough.