ImGuiDiskManipulator: auto refresh msx-side
Every frame refresh the MSX side. That is: do a "DIR" of the MSX files,
and re-sort.

This didn't noticably increase CPU usage (much less than 1%). And having
a guaranteed up-to-date view is certainly worth this (small) cost.

And this allowed to remove the "manual refresh" button from th MSX side.