Remove frame_rate_pusher Tcl script
This only did something for Android: it forces a screen refresh at a
rate of 25fps by constantly moving an invisible OSD element. (See
comments in the now removed script for the motivation).

In the past this was useful, because when nothing changes on the screen
(e.g. the MSX is paused) then we wouldn't redraw the screen. Though now
that we're using Dear ImGui we do already force a minimal framerate of
25fps.