ImGuiDiskManipulator: confirm before overwriting MSX files
Before this patch already existing files would be silently(!) preserved.
Now you get a question whether to overwrite or preserve the files (or
directories).

We also detect that two or more host files will map to the same MSX
filename (in 8.3 filename format). And if so, ask for confirmation.

With this patch all the ideas I had to improve ImGuiDiskManipulator are
implemented. But further ideas are always welcome of course.