ImGui: added 'Test MSX hardware' window
To keep openMSX responsive (don't block for several seconds while the
window is opening) we only test 1 machine or extension per iteration.
That is one per frame (typically 50 or 60 per second).

Currently that takes about 7 seconds (but only the first time this
window gets opened this session). We could easily make this 2x, 3x, 4x,
... faster. Though seeing the counters go up is also a nice effect ;-)

I also added a button to copy the list of non-working hardware, together
with the reason why, to the clipboard.