Simplify usage of ImGuiListClipper
"ImGuiListClipper" is a utility to only draw that portion on a (long)
list which is actually visible. Typically this is used in a ListBox,
ComboBox, Table, ... This is great to improve efficiency, though using
it correctly isn't trivial:

    ImGuiListClipper clipper;
    clipper.Begin(narrow<int>(myList.size()));
    while (clipper.Step()) {
        for (int i : xrange(clipper.DisplayStart, clipper.DisplayEnd)) {
            draw_stuff(myList[i]);
        }
    }

This patch adds a utility im::ListClipper() that allows to simplify this
to:

    im::ListClipper(myList.size(), [&](int i) {
        draw_stuff(myList[i]);
    }

This removes 4 lines of code for each usage, but more importantly it
reduces the indentation level (here the body is just a single line, in
actual code it's often much longer).

To be honest: im::ListClipper only covers the most typical use case of
ImGuiListClipper. But of course for the more complex use cases you can
still use ImGuiListClipper directly.