Attempt to test extensions in incomplete openMSX installation
Testing extensions requires a MSX machine instance. We use the
'C-BIOS_MSX1' machine for this, because it's a small and most extensions
work in this machine. It should also be present in all openMSX
installations ...

... except when it is not, meaning the openMSX installation is
incomplete. Normally this shouldn't happen, but for openMSX developers
(depending on how you organize your ~/.openMSX directory, e.g. as a link
to the openMSX git repo) it can happen they are missing afterall.

This patch does a minimal attempt to recover from this situation: it
tries to use the (a different instance of) the current MSX machine to
test the extensions.