Tweak calls to ImGui::TextUnformatted()
Change:
    ImGui::TextUnformatted("abc");
to:
    ImGui::TextUnformatted("abc"sv);

These respectively expand to
    ImGui::TextUnformatted("abc", nullptr);
    ImGui::TextUnformatted("abc", 3);

The former calculates the length of the string literal at run-time, the
latter does this calculation at compile-time.

The "sv" suffix is ugly. Usually you avoid this by not offer an overload
that takes a "const char*" argument (but instead one that takes a
std::string_view argument). But in this case the function is part of the
"Dear ImGui" library, and we can only add additional overloads, not
remove overloads. (We added an overload that takes a std::string_view).