ImGuiMedia: tweak drawing separator lines in menu
In "Main menu > Media" we draw horizontal separator lines between the
sections:
  cartridge slots
  extensions
  disk drives
  tape deck
  harddisk
  cdrom
  laserdisc

Some of these sections can be empty. And then we don't want to draw two
consecutive separator lines. We only want to draw separator lines
between sections, so no line after the last section.

This patch fixes a bug were we did draw a line after the 'tape deck'
section when all following sections were empty.

This used to work fine. But we introduced a bug when we started to
replace (some of the) empty sections with (grayed out) messages like:
  No cartridge slots present
  No disk drives present
  No cassette port present