ImGuiConsole: fix drawing (colored) non-ascii characters
Because Dear ImGui does not (yet) support colored text in an InputText
field (for syntax highlighting in the console) we have a hack were we
overdraw the white text afterwards in the correct colors.

But there was a bug: some calculation assumed each character takes a
single byte. But that's of course wrong for non-ascii UTF-8 text.

Easy fix once the problem was identified.