Raise (some) windows when opening them
Most windows are opened/closed via a check mark. For example 'Tools >
Disk manipulator' toggles the 'disk manipulator' window open/close.

Some windows have an associated button that always opens the window (no
toggle). For example the 'Open palette editor' button in the bitmap
viewer, or the 'Configure...' button in the 'Message log' window (to
open the 'Configure messages' window).

For the latter group of windows it was possible that the window was
already open, but hidden behind some other window. If you then press the
button nothing happened (because the window was already open). This
patch will now also raise the window. And this is similar behavior as
what you get for the former group: toggling the window closed and then
open again will always show the window on top.

This addresses issue:
    https://github.com/openMSX/openMSX/issues/1557