(Try to) replace all 16-bit constants with labels in disassembly view
Before we only replaced addresses, for example in instructions like:
  jp nn
  ld hl,(nn)
But not immediate values (which may or may not be addresses) in
instructions like:
  ld de,nn
After this patch we'll now also replace the latter.

We do not try to replace 8-bit immediate values in instruction like
  ld r,n

Note that these are all only heuristics. Without access to the original
source code there's no way to 100% correctly substitute labels. E.g.
there may be multiple labels with the same numeric value. Or there may
be a (unique) label with a matching numeric value, but the original
source code still used an immediate value rather than this label. Etc

This was discussed in:
    https://github.com/openMSX/openMSX/issues/1553