ImGuiDebugger: show popup in register view
Which shows the register value formatted as binary and decimal (the
regular view only shows hexadecimal). And for 16-bit values, also show
the decimal value of the high and low 8-bit parts.

Part of this patch extends our strCat() utility with:
* bin_string<N>(val): format value as binary with exactly N digits
* dec_string<N>(val): format as decimal with leading spaces
Similar to the existing 'hex_string<N>(val)' functionality.
And extends the unittest for this.