Don't hardcode imgui color values
Instead always use either:
* named Dear ImGui colors, e.g. ImGui::GetColorU32(ImGuiCol_ButtonHovered()
  (this is an existing mechanism)
* named openMSX imgui colors e.g. getColor(imColor::ERROR)
  (this a new mechanism introduced in this patch)

TODO fine-tune the colors in ImGuiUtils.cc setColors() (around line 260)
depending on the selected color style (light or dark).

This change was triggered by this issue:
    https://github.com/openMSX/openMSX/issues/1567
Thanks gilbertfrancois for reporting!

Maybe not everything is fixed already, but at least now we have the
infrastructure to address it.