Save/restore position of main openMSX window
Dear ImGui already saves/restore the position of all other windows. It's
unexpected that the position of the main openMSX window was not. The
underlying reason is that Dear ImGui does not manage the main window,
instead we do that via SDL2.

This patch is complex (as is: a lot of code, not difficult code) for
these reasons:
* It's not always possible to query the position of the main window:
  * In full screen we don't want to get position (0, 0) instead we want
    to remember the position in windowed mode.
  * The none-renderer doesn't have any window.
  * While switch scale-factor we first destroy the window, and only then
    recreate the new window.
* ImGuiManager is the most logical place to save the window position (in
  the imgui.ini file). For technical reasons 'ImGuiManager' must have a
  longer lieftime than 'Display'. Thus during saving (on exit) we can no
  longer query the window position.

Because of the combination of the above, we store the 'last known'
window position in ImGuiManager. And then 'ImGuiManager' and 'Display'
coordinate to keep that information up-to-date, or retrieve it when
(re)creating the window.

The classes 'Display', 'VideoSystem' and 'VisibleSurface' coordinate to
actually query the position of the main window (when present) or change
the position.