Draw first frame before starting emulation
This addresses the issue:
    ImGui Debugger: breakpoints aren't saved between sessions like with
    the old debugger. The breakpoint list still shows the break points,
    but there is no break done. To reproduce: create a breakpoint on
    address 1 and boot the MSX. It will break. Then stop openMSX and
    start openMSX. The breakpoint still shows on 1, but no break is
    executed.

Actually the breakpoints _were_ correctly restored, only too late:

Breakpoints are stored in the GUI session (the imgui.ini file). This
file is parsed on the first Dear ImGui frame, so also on the first
openMSX frame. Before this patch we would draw the first frame when the
VDP has finished rendering its first frame. So typically 1/50 or 1/60 of
a second _after_ emulation has started running.

But in this context that means the breakpoints are only restored a
little while after the start of emulation. And thus a breakpoint near
the start of emulation indeed didn't trigger.

Fixed by rendering a frame (a black frame) right before we start
emulating.