Use unique_ptr to improve incremental compilation speed
Before this patch in ImGuiManager we stored a lot of 'ImGuiXXX' objects
by value. To do that we needed to #include the corresponding headers in
'ImGuiManager.hh'. 'ImGuiManager.hh' itself was #included in many other
files. So very often when you change something in ImGui-related code
(almost) all of the other ImGui code also needed to be recompiled.

This patch stores those objects by 'unique_ptr'. This only requires a
forward declaration of the class. And then we can move the #include from
the .hh to the .cc file.

It's a trade-off:
* it's a tiny bit slower (probably not measurable)
versus:
* faster incremental compilation speed (when working on ImGui stuff).