Fix crash when file for the currently inserted rom is deleted
Reported by VampierMSX, thanks a lot!

Steps to reproduce:
* Start openMSX with a ROM file.
* While openMSX is running, remove that ROM file.
* Drag and drop a different ROM file on the main openMSX window
--> openMSX crashes with:
    Uncaught exception: File "..." not found

Problem was located indirectly in the ImGui drag-and-drop handling code
where we try to obtain the "display name" for the ROM. Internally this
uses the sha1 of the file to look it up in the database and then get the
ROM title. But this didn't handle the case that the file went missing
(thus calculating the sha1 is no longer possible).

Easy fix once the problem was identified.

I also did an audit of our code for all callers of
    Sha1Sum FilePool::getSha1Sum(const std::string& filename);
Some already handled missing files (by catching an exception), others did
not (and then possibly crash). To make the code more robust I changed
the above function signature to:
    std::optional<Sha1Sum> FilePool::getSha1Sum(const std::string& filename);
Now all callers are forced to deal with a possible inability to
calculate the sha1.