ImGuiSettings: Fix possible crash when changing settings
E.g. in 'Settings > Advanced > All settings', look for the
'color_matrix' settings and edit the string into something invalid (e.g.
I added a comma at the end).

Thus used to crash (uncaught exception). Now this exception is caught,
and then ignored. The GUI then reverts to the last valid value for the
setting. Maybe in some cases an error message would be better, in other
cases not. But in any case this is already a lot better than crashing.