ImGui: don't focus reverse-bar / OSD-icons on startup
By default in Dear ImGui, when a window first appears it gets focus.
Usually this is what you want. Though with the default settings, when
openMSX starts, it creates a window for the reverse-bar and the
OSD-icons. And thus one of these windows got focus instead of the main
openMSX window. Because of that you first had to click in the openMSX
window (to give it focus) before you e.g. could start typing in the MSX.