Change font 'Vera' to 'DejaVu'
It should be a strict superset, see here for some history:
  https://en.wikipedia.org/wiki/DejaVu_fonts

The motivation for this change was that the character (glyph) for U+03BC
(greek small letter mu) is missing from the 'Vera' font (it is present
in 'DejaVu'), even though the visually similar character for U+00B5
(micro sign) is present. And this character is used in one of the
extensions that we ship (mu-PACK.xml).