ImGuiDebugger: added 'run-to' functionality in context menu
Triggered by:
    https://www.msx.org/forum/msx-talk/openmsx/please-help-testing-upcoming-openmsx-release?page=77
    Post by Bengalack on 03-01-2024, 19:44

The old debugger had a "Run to" button in the toolbar. To use it, you
first had to click on a line in the disassembly view, then click on this
button.

This patch adds a "Run to" option in the context menu of the disassembly
view. So to use it: right click on a line in the disassembly view, then
in the context menu select "Run to". IMHO that's more logical than
having the functionality spread over two widgets.

Next to "Run to <the selected line>", this patch also adds a "Run to
<edit field>". This allows to run to any address (or symbol value),
without first having to scroll to that address.