Fix possible crash with replay and watchpoints
Consider this scenario:
* debug set_watchpoint write_mem 0x8000 1 { poke 0x8000 1 }
  This means: after every write to address 0x8000, overwrite that
  address again with the value '1'. Effectively turning this memory
  location into ROM. (I'm simplifying a bit).
* Trigger this watchpoint (e.g. by executing POKE in BASIC).
* Then replay this scenario (e.g. by pressing PgUp a few times).
Before this patch this caused a crash.

The reason is that the Tcl command 'poke ..' is a 'RecordedCommand'.
This means it's a command that should be included in the replay (because
it changes the state of the MSX machine). During a replay we re-execute
the 'poke' command, and this also re-triggers the watchpoint, which
triggers another 'poke' command. But here comes the problem:

When a new state-changing action is executing we currently stop replay.
Usually that's intended, e.g. during replay you give new keyboard input.
But in this case the new action (=the command triggered by the
watchpoint) is triggered from the replay itself. This has two problems:
* A replay shouldn't interrupt itself.
* The internal bookkeeping didn't expect this scenario, and that then
  triggered a crash (an assert).

My first idea for how to solve this was to not execute debugger
callbacks (on watchpoints, breakpoints, probes, ...) during replay.
Though this possibly excludes some useful scenarios. E.g. you might be
debugging a replay, and then you really want to be able to use
breakpoints during replay. Or you might have a logging system based on
watchpoints. So this idea doesn't work.

Instead this patch only disables executing _new_ 'RecordedCommands'
during replay that get triggered by executing of debugger callbacks. So:
* Previously recorded commands (like 'poke') still get replayed.
* New recorded commands outside of debugger callbacks still interrupt
  replay (e.g. a manual 'poke' command in the console during replay).
* Non-recorded commands triggered from debug callback during replay (e.g.
  'debug break', or some kind of logging) still execute.