Fix order of write_mem-callback replay
This is a follow-up on the previous patch. There we had the scenario:
* debug set_watchpoint write_mem 0x8000 1 { poke 0x8000 1 }
* trigger this watchpoint
* and replay it.

Depending on various conditions this replay could go wrong. More
specifically: the order of the initiator of the watchpoint (a memory
write) and the triggered callback command could be reversed.

For a memory-write-watchpoint:
* We first do the actual memory write at a specific moment in time 't'.
* Then we execute the Tcl command associated with this watchpoint. This
  command executes at the same time 't'.
* If this is a 'RecordedCommand', the replay-log contains this command
  also at time 't'.

Often this is the desired order (first do the write, then trigger the
callback). E.g. this allows to undo the write in the callback. Note that
for read-watchpoints the order is reversed: first callback, then the
actual read (e.g. this allows for the callback to change what will be
read).

During replay:
* We replay the recorded commands at their specified timestamps.
* We also replay the CPU instructions and the memory read/writes
  associated with those instructions at the correct time.

But there's a mismatch when there's a memory transaction and a
replay-event that are both associated with the same timestamp. Currently
such simultaneous events are resolved by first replaying the event and
only then executing the (partial) instruction.

More in detail:
* During CPU emulation, right before every 'communication' (IO access or
  memory access to a memory mapped device) we first check for (and
  execute) 'synchronization points'.
* Replay events are scheduled via such 'synchronization points'.
* Thus they now execute right before the transaction that triggered
  them.

And usually that's the correct order. But not in case of a write-memory
watchpoint callback (note: in case of a read-watchpoint it's fine).

As a solution this patch executes the write-callback a timestamp that's
a tiny bit later than the timestamp of the write transaction.

--

There's one more complication, and that's not yet solved in this patch.

Above I said that before every 'communication' the CPU emulation checks
for synchronization points. Such a check is not expensive, but CPU
emulation must be very fast. The CPU does literally a million
'communications' per second. And then doing a check before _all_ of them
does become expensive. We've solved that by excluding "backdoor
communication" from the "syn-point-check".

"Backdoor communication" are memory read/writes to plain ROM/RAM
regions. These are by far the most common, so it makes sense to optimize
those. Plain ROM/RAM read/writes don't trigger any side-effects in
devices. So it's fine (=correct) to not trigger syn-points exactly at
the right moment in time. Instead, when only backdoor access is
happening, synchronization is only accurate to whole instruction
boundaries, not sub-instruction accuracy. And again: that's fine.

When read or write watchpoints are attached to a memory region those
memory regions are excluded from "backdoor" accesses. And then
everything is again fine. In particular: watchpoint actions and
callbacks all trigger in the correct order.

The problem occurs when replaying such an action and then more
specifically when during replay the memory region has no watchpoint
(anymore). In that case the sync-point only triggers with instruction
(not sub-instruction) accuracy. Thus it's possible that during replay
the full instruction finishes and only then the watchpoint callback
triggers. While in the original recording first the callback executes
and only then the instruction finishes.

Note that this only matters for read_mem watchpoints (not write_mem, nor
IO watchpoints). At the moment I don't know of any practical scenarios
where this turns out to be a problem. So, at least for now, I'll keep
this (theoretical?) problem unsolved. (Because ATM I don't see an easy
and fast solution yet).