[0/n]: Move imgui_memory_editor.h to DebuggableEditor.hh
This patch series makes several changes/improvements to the
imgui_memory_editor widget. This is a 3rdparty widget (license: MIT).
* The upstream version hasn't been changed in a long time.
* It's not too much code (~700 lines).
* We will make a lot of changes that are fairly specific to openMSX (not
  easy to generalize and push upstream).

Because of the combination of the above reasons, we will copy this
widget into our (part of the) source tree, and start modifying it.

The only thing this patch does is move that code, without making any
other changes yet.