[1/n] Adjust coding style
Because this widget is now in "our" source tree, and we plan to make
several code changes, it makes sense to also rewrite it (more) into our
coding style.

This patch makes the bulk of the code-style changes, without altering
any of the functionality yet. More (but much smaller and more localized)
style changes will follow in later patches.