[2/n] remove 'OptShowHexII'
We want to change/extend the (former) imgui_memory_editor.h code. That's
easier if there's less code to change (and less code to understand). So
first we'll remove soem code that we don't really need.

This patch removes the 'Show HexII' option. I personally think it's a
weird option. I haven't seen this is (m)any other hex-editor(s). So I
don't think many people will miss this.

Also part of the motivation to remove this is that it doesn't interact
well with the ability to edit the ascii-side. Currently that's not
supported, but it's something that we do want eventually (but it's not
yet part of this patch series).