[4/n] remove custom highlight-range / float/double-preview
It was possible (only via code) to set a custom hightlight region. We
don't need that functionality (and it complicates code that we want to
change). So remove it.

In the data-preview-window it's possible to select the data-type. his
could be various integer types (8,16,32,64 bit / (un)signed /
big/little-endian). But also IEEE 'float' and 'double'. The latter two
don't make sense in the context of MSX (MSX-BASIC does have floating
point values, but those use a totally different representation than
IEEE). So remove these two types.