[6/n]: simplify highlight logic
The original code could highlight regions for 3 reasons:
* Via a custom region defined via upper/lower boundary in code. This was
  already removed in a previous patch in this series.
* It highlights the preview data-type (only visible when that type is
  wider than 1 byte). But only when the data-preview window is actually
  open.
* There's a highlight callback function.

In openMSX we used the callback function to re-implement the preview
highlight logic, but now in a way that it's always active, also when the
preview window is not open. Without this extra highlight, the cursor
position was not visible when the hex-editor-window did not have focus.

But now that we're changing the code, we no longer need to augment the
highlight logic via a callback, we can directly change it in the code.
This allows to greatly simplify the code:
* Remove the callback functionality.
* No longer re-implement the preview highlight logic, just _remove_ the
  check that preview is active.

While working on this, I noticed that 'DataPreviewAddr' and
'DataEditingAddr' are always in sync. So no need to have both.

I changed the default preview-data-type from signed-32bit-integer to
unsigned-8bit-integer. That's a better default in the context of MSX.