[7/n] use 'unsigned' for address
The original code used a 'size_t' variable (64-bit on modern platforms)
to store an address. That makes perfect sense.

In the context of openMSX, 'debuggables' only use a 32-bit address (more
than sufficient for any MSX device).

Printing 'size_t' variables via printf() is different in Linux and
Windows. The code used a macro to handle this difference.

Because of the combination of the above I decided to just use a 32-bit
variable to store addresses in thise code. And as a bonus that allowed
to remove some casts when interfacing with other openMSX code.