ImGuiTrainer: fix empty list when openMSX is started with this window open
On first use ImGuiTrainer queries the list of available trainers. Though
when this window is already open when openMSX starts (because it was
left open in the previous session), this query can get executed before
all startup Tcl scripts have been executed. Then the query fails, and
the list of available trainers is permanently empty. (We intentionally
don't keep on retrying).

Fixed by only doing this query after everything is fully started.