[9/n] split in .hh and .cc / base-class / ReadFn+WriteFn
Till now the hex-editor widget was fully implented in the .hh file. This
patch splits it over a .hh and a .cc file.

We used the 'MemoryEditor' class as a base class to implement
'DebuggableEditor'. This derived class does basically nothing but
* setup the initial state (ReadFn and WriteFn callbacks).
* adapt some parameters and delegate to the base class.

Now that we are modifying the code, we can skip this intermediate step:
* We integrate 'DebuggableEditor' into 'MemoryEditor' and rename the
  class to 'DebuggableEditor'.
* We change the interface to directly work with openMSX 'Debuggables'.
* And because of this we no longer need the ReadFn and WriteFn callback
  functions.