[12/n] Allow to freely resize window
Before this patch the hex-editor window had a fixed width (the height
could be resized). This width was set by the number of columns and
whether ascii-view was shown or not. After this patch these things only
set the initial width, but afterwards it can be freely resized.

Note that making the window narrower/wider does not automatically
decrease/increase the number of columns (instead you now get a
horizontal scrollbar when the window is too narrow). This new behavior
fits better when the hex-editor is docked in a larger window. Manually
specifying the number of columns (possibly with a scrollbar) is usually
better than having a 'weird' number of columns.

I was testing the hex-viewer on the 'MapperIO' debuggable. This
debuggable has a size of only 4 addresses:
* By default we show 16 columns, but that looks weird when the total
  size is less than that. So now we limit the maximum number of columns
  to the size of the debuggable.
* When the full debuggable can be shown on a single line, the address
  bar is not so useful (and it even takes up more space than the small
  hex-data itself). So now(*1) we allow to hide the address bar(*2).
  This is especially useful if you want to dock many small hex-editors in
  a larger window (without wasting too much space).

(*1) The address bar could already be hidden in the original code, we
removed that possibility in commit 5ad9b09a50. Though one major
difference is that before it could only be controlled from c++ code. Now
the user can make this choice.

(*2) For this option to be useful it would really be nice if it were
saved over openMSX sessions (and saved per hex-editor instance). That
might be one of the next things I'll work on.