[1/5] Split ImGuiPart in two ...
... in ImGuiPart annd ImGuiPartInterface.

This is a preparation to save the settings of the (indivudual) hex-
editors. For this we need to be able to dynamically create 'ImGuiPart'
objects, and (un)register them in the ImGuiManager class. Though before
this patch ImGuiManager is itself a ImGuiPart object. And
(automatically) registering an object into itself runs into lifetime
problems. Those problems disappear when we split ImGuiPart in two.

This patch also does some cleanups: because ImGuiPart now stores a
reference to ImGuiManager, we can remove that reference from many other
classes.