[4/5] refactor DebuggableEditor::paint()
Give it the same interface as ImGuiPart::paint() (as preparation to
inherit from ImGuiPart).

This requires storing the name of the debuggable (and also the window
title (*1)) in the DebuggableEditor object. So now we pass this
information to the constructor and store it.

(*1) When there are multiple DebuggableEditors for the same debuggable,
these editors have the same debuggable name, but a different window
title.