[5/5] save settings for DebuggableEditor
The previous patches in this series were (mostly) preparations to make
DebuggableEditor inherit from ImGuiPart. This patch adds that
inheritance.

The main motivation for this inheritance is that ImGuiPart objects can
save/restore their state in the "imgui.ini" file. And then this patch
also implements save/restore for DebuggableEditor.

Implementing save/restore is often simply enumerating which member
variables should be included. The only extra thing we needed to do here
was extend the save/restore infrastructure with support for two new
types ('unsigned' and 'std::string').