ImGuiSettings: some colors could be wrong (=black) the first time openMSX starts
More specifically when openMSX starts without an "imgui.ini" file. Or
when that file does not contain a "[openmsx][settings] style=<x>" line
(should only happen when that file was manually edited).

When no style is explicitly loaded from "imgui.ini" we keep the default
Dear-ImGui style. But then we forgot to initialize the extra openMSX
style-colors (e.g. used for the openMSX console).