DebuggableEditor: fix update value under cursor ..
.. when that cell loses focus (e.g. by right-clicking, or by clicking in
another window).

This functionality got broken during refactoring (in commit 332ab13696).

This patch also draws the cursor in the ascii-part in grey (as opposed
to blue) to more clearly indicate that keyboard input influences the
hex-party (which is drawn in blue when active). In the future we could
use the blue/grey colors to indicate which part accepts input (though
atm I'm not yet actively working towards this goal).