ImGuiBreakPoints: clear breakpoints on load-settings
See here for the bug report:
    https://www.msx.org/forum/msx-talk/openmsx/please-help-testing-upcoming-openmsx-release?page=83
    Post on '29-01-2024 23:44' by 'DamnedAngel'
Breakpoints were _added_ to the existing breakpoints instead of
_replacing_ them. (And same for watchpoints and debug conditions).

I also added a 'Clear' button (below the 'Add' and 'Remove' buttons) to
remove all breakpoints. Fixing the above required implementing this
functionality. So it was only a small extra step to expose it in the
GUI.