Remove Joystick class
Some time (multiple months) ago we added 'MSXJoystick' this was a more
flexible replacement for both the (already removed) "keyjoystick{1,2}"
pluggable and for the "joystick<N>" pluggables. This patch now removes
the latter.

The motivation for removing this now is that we want to "fix" how
openMSX interacts with the SDL joystick API (see later patches for more
details). And refactoring is always easier when there's less code to
worry about.

This patch (temporarily) breaks all joystick stuff (as in: host
joysticks don't respond anymore) because there's no code anymore that
calls SDL_JoystickOpen(). This will be restore in the next patch.

--

This patch also removes all "#ifndef SDL_JOYSTICK_DISABLED" lines.
Meaning we now require an SDL version which includes joystick support.

Actually we already had this requirement for some time: we did call
several SDL joystick functions without this #ifdef protection. Nobody
complained, so nobody is trying to use openMSX in combination with an
SDL version without joystick support.

For now this simplifies the code. If we ever need this again in the
future (I doubt we will), it shouldn't be too hard to re-add it.