Handle SDL2 joystick add/remove events
OpenMSX joystick support was added when we were still using SDL1. SDL1
only scans the host joysticks at startup. SDL2 can handle plugging /
unplugging of host joysticks while the application is running. OpenMSX
was never adapted to take this possibility int account (... until now).

The obvious problem with this is that when you plug a new host joystick,
openMSX won't be able to use it. And from an SDL point of view
unplugging and replugging the same host joystick is also seen as a new
host joystick.

A second problem with this is that querying information from a no longer
present host joystick can result in a crash. Since recently this became
possible (since ImGuiSettings wanted to show the host joystick name).

--

In short, SDL2 enumerates the host joysticks in 2 ways:
* Via a device index: these are always 'dense' numbers in the range
  0..SDL_NumJoysticks(). These values should only be used when
  'discovering' host joysticks. Either when enumerating the currently
  present joysticks, or the SDL_JOYDEVICEADDED event also carries a
  device index.
* Via an instance ID: these are unique numbers given to each new host
  joysticks. A newly plugged host joystick gets a new instance ID.
  Instance IDs are never reused within a session. For example SDL
  joystick input events carry an instance ID.
Before this patch openMSX was wrongly using device indices and instance
IDs as if there were the same thing (as was the case with SDL1).

This patch introduces a 3rd, openMSX specific, way to enumerate host
joysticks: openmsx::JoystickId

This is a _minimal_ attempt to give stable IDs to host joysticks.
* Suppose you have 2 host joysticks plugged when openMSX starts.
* These joysticks will have openMSX joystick names 'joy1' and 'joy2'
  (internally these have numeric values 0 and 1).
* When you unplug the joystick that corresponds to 'joy1'. That name
  will no longer be available, but 'joy2' remains unchanged (e.g.
  bindings to 'joy2' remain associated with the same host joystick).
* When a new host joystick is plugged (this could be either the
  previously unplugged joystick or another one) this new host joystick
  will recycle the name 'joy1'.
* More in general:
  ** Whenever a new host joystick becomes available it will be given the
     lowest available free name 'joyN'.
  ** Whenever a host joystick gets removed, the associated new becomes
     free, but it doesn't affect the openMSX names of any other
     joysticks.

Note that this is only a _minimal_ solution:
* It does not guarantee that when a joystick is unplugged and then
  immediately replugged, it retains the same name.
* It does not guarantee that host joysticks (when more than one is
  present) get the same names over different openMSX sessions.

--

This patch adds a new class: JoystickManager:
* This manages the mapping between SDL instance IDs and openMSX
  JoystickId's (done by monitoring SDL_JOYDEVICE{ADDED,REMOVED} events).
* All direct interaction with the SDL2 joystick API goes via this class.
  All other openMSX code only indirectly interacts with host joysticks
  via joystick events.
* It creates the 'joystick<N>_deadzone' settings. It will dynamically
  create more settings when needed.

--

An important design decision of this patch is to represent openMSX
JoystickIds via a separate class instead of via an integer (even though
JoystickId only contains a single integer as member variables).

And JoystickId intentionally only offers a minimal API. In particular it
does NOT implicitly convert to an integer, and it does not support
support arithmetic.

This was it was relatively to easy to change some function to return a
JoystickID (instead of an integer) and then (via error messages) the
compiler helps a lot in catching all places that require attention for
converting the old SDL IDs to the new openMSX JoystickIDs.

It also eliminates off-by-one errors in converting IDs to string
representation (ID=0 corresponds to the string "joy1"). That logic is
now located in exactly 2 places: 1 place to convert ID to string and
1 to parse a string to an ID.